#  Enforce page alignment for sampled allocations