This allows sampled allocations to be checked through alignment, therefore enable sized deallocation regardless of cache_oblivious.

Enables the speed up of sized deallocation, in `test/integration/stress`:
before:

    100000000 iterations, dallocx=1254009us, sdallocx=1222009us, ratio=1:1.026

with this PR:

    100000000 iterations, dallocx=1259009us, sdallocx=759005us, ratio=1:1.659

Resolves #1697